refactor(result): use named import { ProductCard } and unify path
