+++ components/result/FirstPickSingle.tsx
@@ -1,4 +1,4 @@
-import ProductCard from "@/components/result/ProductCard";
+import { ProductCard } from "@/components/ProductCard";
 
 export default function FirstPickSingle({ product }: { product: any | null }) {
   if (!product) return null;
